Major updates
Provides options in the plugins admin area around what langs are supported and offers a choice of skin.
Reinstates line-numbering.
Supports <code> and <pre> tags with the prettyprint class (but not nested <pre><code></code></pre> which is supported by prettyprint)
Moves javascript inclusion to the footer and loads additional resources async
Only loads prettyprint JS if matching code blocks are found in the content